update valid_enum_value parameter type and improve oneof_index check

+++ include/hpp_proto/descriptor_pool.hpp
@@ -282,9 +282,9 @@ public:
 
     [[nodiscard]] bool is_closed() const { return options_.features.value().enum_type == FeatureSet::EnumType::CLOSED; }
 
-    [[nodiscard]] bool valid_enum_value(uint32_t v) const {
+    [[nodiscard]] bool valid_enum_value(int32_t v) const {
       return !is_closed() ||
-             std::ranges::contains(proto_.value, static_cast<int32_t>(v), [](const auto &item) { return item.number; });
+             std::ranges::contains(proto_.value, v, [](const auto &item) { return item.number; });
     }
 
   private:
